+++ chaosmagpy/__init__.py
@@ -1,3 +1,25 @@
+"""
+`chaosmagpy` is a simple Python package for evaluating the CHAOS geomagnetic
+field model and other models of Earth's magnetic field. The CHAOS-7 model can
+be downloaded at `http://www.spacecenter.dk/files/magnetic-models/CHAOS-7/`.
+
+>>> import chaosmagpy as cp
+>>> chaos = cp.load_CHAOS_matfile('CHAOS-7.mat')
+
+The following modules are available:
+
+* chaosmagpy.chaos
+* chaosmagpy.coordinate_utils
+* chaosmagpy.model_utils
+* chaosmagpy.plot_utils
+* chaosmagpy.data_utils
+* chaosmagpy.config_utils
+
+More information on how to use ChaosMagPy can be found in the documentation
+available at `https://chaosmagpy.readthedocs.io/en/master/`.
+
+"""
+
 __all__ = [
     "CHAOS", "load_CHAOS_matfile", "load_CHAOS_shcfile",
     "load_CovObs_txtfile", "load_gufm1_txtfile", "load_CALS7K_txtfile",
@@ -5,7 +27,7 @@ __all__ = [
     "mjd_to_dyear"
 ]
 
-__version__ = "0.11"
+__version__ = "0.12-dev"
 
 from .chaos import (
     CHAOS,